+++ drizzle/0017_experimental_wechat.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `settings` ADD `experimental_wechat` integer DEFAULT false NOT NULL;